+++ R/regional_rt_pipeline.R
@@ -9,6 +9,7 @@
 #' @param national Logical defaults to `FALSE`. Should a national summary nowcast be made.
 #' @param regional_delay Logical defaults to `FALSE`. Should reporting delays be estimated by region.
 #' @param merge_onsets Logical defaults to `FALSE`. Should available onset data be used. Typically if `regional_delay` is
+#' @param case_limit Numeric, the minimum number of cases in a region required for that region to be evaluated. Defaults to 20.
 #' set to `FALSE` this should also be `FALSE`
 #' @param ... 
 #' @inheritParams rt_pipeline
@@ -16,19 +17,33 @@
 #' @export
 #' @importFrom furrr future_map
 #' @importFrom tidyr complete
-#' @importFrom dplyr count filter rename
+#' @importFrom dplyr count filter rename filter group_by pull
 #' @examples
 #' 
 #' ## Code
 #' regional_rt_pipeline
 regional_rt_pipeline <- function(cases = NULL, linelist = NULL, target_folder = "results", 
-                                 national = FALSE, regional_delay = FALSE, merge_onsets = FALSE, samples = 1000, ...) {
+                                 national = FALSE, regional_delay = FALSE, merge_onsets = FALSE,
+                                 case_limit = 20,
+                                 samples = 1000, ...) {
   
   
   ## Control parameters
   target_date <- as.character(max(cases$date))
   
   message("Running pipeline for ", target_date)
+  
+  
+  ## Check for regions with fewer than required cases
+  zero_regions <- cases %>% 
+    dplyr::group_by(region) %>% 
+    dplyr::count(wt = cases) %>% 
+    dplyr::filter(n < case_limit) %>% 
+    dplyr::pull(region)
+  
+  ## Exclude zero regions
+  cases <- cases %>% 
+    dplyr::filter(!region %in% zero_regions)
   
   ## Make sure all dates have cases numbers
   cases <- cases %>% 
@@ -49,7 +64,7 @@ regional_rt_pipeline <- function(cases = NULL, linelist = NULL, target_folder = 
     rt_pipeline(
       cases = national_cases,
       linelist = linelist,
-      target_folder = file.path(target_folder, "national", target_date),
+      target_folder = file.path(target_folder, "national"),
       target_date = target_date, 
       merge_actual_onsets = merge_onsets, 
       samples = samples, ...)
@@ -83,7 +98,7 @@ regional_rt_pipeline <- function(cases = NULL, linelist = NULL, target_folder = 
     rt_pipeline(
       cases = regional_cases,
       linelist = regional_linelist,
-      target_folder = file.path(target_folder, target_region, target_date),
+      target_folder = file.path(target_folder, target_region),
       target_date = target_date, 
       merge_actual_onsets = merge_onsets, 
       samples = samples, ...)