+++ NEWS.md
@@ -1,6 +1,8 @@
 # EpiNow 0.1.0
 
 * Added a `NEWS.md` file to track changes to the package.
-* Extract tooling from global nowcasts
+* Extract tooling from global nowcasts.
 * Parameterised the report to be generalised.
-* Added a regional pipeline and made a national summary nowcast optional
+* Added a regional pipeline and made a national summary nowcast optional.
+* Added summary and plotting functions built to work from saved results.
+* Added a function to summarise results across regions